ActiveJob refactor
