+++ app/models/spree/product_import.rb
@@ -15,6 +15,9 @@ module Spree
     has_attached_file :data_file,
       path: ":rails_root/tmp/product_data/data-files/:basename_:timestamp.:extension",
       url: ":rails_root/tmp/product_data/data-files/:basename_:timestamp.:extension"
+
+    belongs_to :user, class_name: 'Spree::User', foreign_key: 'created_by', inverse_of: :product_imports
+
     validates_attachment_presence :data_file
     #Content type of csv vary in different browsers.
     validates_attachment :data_file, :presence => true, content_type: { content_type: ["text/csv", "text/plain", "text/comma-separated-values", "application/octet-stream", "application/vnd.ms-excel", "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"] }
