+++ lib/solidus_import_products/engine.rb
@@ -15,6 +15,10 @@ module SolidusImportProducts
 
       Spree::UserMailer.send(:include, SolidusImportProducts::UserMailerExt)
 
+      # TODO: replace the class_eval
+      Spree::User.class_eval do
+        has_many :product_imports, class_name: 'Spree::ProductImport', foreign_key: 'created_by'
+      end
     end
 
     config.to_prepare &method(:activate).to_proc
