+++ spec/support/active_job.rb
@@ -1,0 +1,9 @@
+ActiveJob::Base.queue_adapter = :test
+
+RSpec.configure do |config|
+  # clean out the queue after each spec
+  config.after(:each) do
+    ActiveJob::Base.queue_adapter.enqueued_jobs = []
+    ActiveJob::Base.queue_adapter.performed_jobs = []
+  end
+end
