--- spec/support/product_import_helpers.rb
@@ -1,8 +1,0 @@
-module ProductImportHelpers
-  def create_import(filename)
-    Spree::ProductImport.create(
-      data_file: File.new(File.join(File.dirname(__FILE__), '..', 'fixtures', filename)),
-      separatorChar: ','
-    )
-  end
-end
